Move test and change docs to clarify transition prioritization in hierarchies
